# lib/httpServer/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Visualize the blockchain by accessing <a href='/blockchain' target='_blank'>here</a>.",
      "version": "0.0.1",
      "title": "Naivecoin API"
    },
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
            "/blockchain/transactions": {
                "get": {
                    "tags": [
                        "blockchain"
                    ],
                    "summary": "Get unconfirmed transactions",
                    "description": "",
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Transaction"
                                }
                            }
                        }
                    }
                },
                "post": {
                    "tags": [
                        "blockchain"
                    ],
                    "summary": "Create a transaction",
                    "description": "",
                    "parameters": [
                        {
                            "in": "body",
                            "name": "body",
                            "description": "Transaction",
                            "required": true,
                            "schema": {
                                "$ref": "#/definitions/Transaction"
                            }
                        }
                    ],
                    "responses": {
                        "201": {
                            "description": "Created",
                            "schema": {
                                "$ref": "#/definitions/Transaction"
                            }
                        },
                        "400": {
                            "description": "Bad Request"
                        },
                        "409": {
                            "description": "Conflict"
                        }
                    }
                }
            },
...
            "/miner/mine": {
                "post": {
                    "tags": [
                        "miner"
                    ],
                    "summary": "Mine a new block",
                    "description": "Mine a new block and add it to the end of the blockchain.",
                    "parameters": [
                        {
                            "in": "body",
                            "name": "body",
                            "description": "Mine request",
                            "required": true,
                            "schema": {
                                "type": "object",
                                "required": [
                                    "rewardAddress"
                                ],
                                "properties": {
                                    "rewardAddress": {
                                        "description": "Address to send the reward",
                                        "type": "string"
                                    },
                                    "feeAddress": {
                                        "description": "Address to send the fee",
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    ],
                    "responses": {
                        "201": {
                            "description": "Created",
                            "schema": {
                                "$ref": "#/definitions/Block"
                            }
                        }
                    }
                }
            }
        },
        "definitions": {
            "Block": {
                "type": "object",
                "properties": {
                    "index": {
                        "type": "integer"
                    },
                    "previousHash": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "integer"
                    },
                    "data": {
                        "type": "object"
                    },
                    "hash": {
                        "type": "string"
                    }
                }
            },
...
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "inputs": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "transaction": {
                                            "type": "string"
                                        },
                                        "index": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "signature": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "outputs": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "amount": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }